refactor(frontend): shared sidebar sections state
Move common code for sections state into `src/lib`.
