+++ frontend/src/data-layers/buildings/state/layer.ts
@@ -2,8 +2,9 @@ import { selector } from 'recoil';
 
 import { ViewLayer } from 'lib/data-map/view-layers';
 import { truthyKeys } from 'lib/helpers';
+import { sectionVisibilityState } from 'lib/state/sections';
+
 import { buildingsViewLayer } from 'data-layers/buildings/buildings-view-layer';
-import { sectionVisibilityState } from 'app/state/sections';
 
 import { buildingSelectionState } from './data-selection';
