+++ frontend/src/data-layers/droughtRisks/state/layer.ts
@@ -1,7 +1,7 @@
 import { selector } from 'recoil';
 
 import { ColorSpec, FieldSpec, ViewLayer } from 'lib/data-map/view-layers';
-import { sectionVisibilityState } from 'app/state/sections';
+import { sectionVisibilityState } from 'lib/state/sections';
 
 import * as droughtRiskColorSpecLookup from '../color-maps';
 import { droughtRiskViewLayer } from '../drought-risk-view-layer';
