+++ frontend/src/data-layers/hazards/state/layer.ts
@@ -3,7 +3,7 @@ import { selector } from 'recoil';
 import { ViewLayer } from 'lib/data-map/view-layers';
 import { truthyKeys } from 'lib/helpers';
 import { dataParamsByGroupState } from 'lib/state/data-params';
-import { sectionVisibilityState } from 'app/state/sections';
+import { sectionVisibilityState } from 'lib/state/sections';
 
 import { hazardVisibilityState } from './hazard-visibility';
 import { HazardParams } from '../domains';
