+++ frontend/src/data-layers/networks/state/layer.ts
@@ -7,10 +7,10 @@ import { recalculateCheckboxStates } from 'lib/controls/checkbox-tree/CheckboxTr
 import { ViewLayer, StyleParams, ColorSpec, FieldSpec } from 'lib/data-map/view-layers';
 import { StateEffect } from 'lib/recoil/state-effects/types';
 import { dataParamsByGroupState } from 'lib/state/data-params';
+import { sectionVisibilityState } from 'lib/state/sections';
 
 import { LayerSpec } from 'app/asset-list/use-sorted-features';
 import { damageMapStyleParamsState } from 'app/state/damage-mapping/damage-style-params';
-import { sectionVisibilityState } from 'app/state/sections';
 
 import adaptationSectorLayers from '../adaptation-sector-layers.json';
 import * as networkColorMaps from '../color-maps';
